Reestructuración y refactorización en obtención de TODOs por usuario
-. Modificamos la query para que únicamente traiga los valores id, description, targetDate, done
-. Para lograr lo anterior, creamos una nueva class TodoSummary en donde no tendremos la variable user_name
-. el ID de la clase pasa a ser un LONG
-. Modificamos las funciones relacionadas a la obtención de TODOS por user y le cambios el type
-. La variable isDone, en la tabla de BBDD se llama DONE, por lo que colocamos nueva anotación @Column cambiandole el nombre
-. En este commit estamos agregando le query para crear un nuevo TODO

+++ src/main/java/com/fullstack/springboot/fsappjavajr/todo/TodoService.java
@@ -30,7 +30,7 @@ public class TodoService {
 		return todo;
 	}
 
-	public void deleteTodoByUserId(int id) {
+	public void deleteTodoByUserId(long id) {
 		Predicate<? super Todo> predicate = todo -> todo.getId() == id;
 		todos.removeIf(predicate);
 	}
